+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All the log of changes on the project/repository
 - Create `minify_img` using the `Imagick` wepb compression
 - Specify libraries and tech-stack base of the proejct in the README.md
 - CLI no action given fallback
+- Create a sitemap file generation following the standard
 
 ### Modified
 - Added "Link to" to `aria-label` in links for a more cohesive screen-reading